packages/bu-tap-twilio/pyproject.toml
[tool.poetry]
name = "bu-tap-twilio"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "bu_tap_twilio"}]

[tool.poetry.dependencies]
python = "^3.10"
singer-python = "^5.13.0"
twilio = "^7.16.1"
python-dotenv = "^0.21.1"

[tool.poetry.dev-dependencies]
pytest = "^7.2.1"
pytest-mock = "^3.10.0"
pytest-cov = "^4.0.0"
pylint = "^2.15.10"
mock = "^5.0.1"

[tool.poetry.scripts]
bu-tap-twilio = "bu_tap_twilio:main"
create-twilio-configs = "bu_tap_twilio.configs:create_twilio_configs"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
